Rearange some stuff for packaging
