+++ src/s2_data/__init__.py
@@ -1,0 +1,2 @@
+from .field_types import BoolType, ByteType, Int32Type, UInt32Type, Int64Type, FlagType, EnumType
+from .field_descriptors import field_descriptors, CharacterUnlockFlags, ShortcutProgressEnum, CharacterSelectedEnum
